Rename modules
* Rename camille.processors -> camille.process
         camille.sources -> camille.source
         camille.utils -> camille util

Move processor tests to their own directory

* We will add source and util sub-packages that will have their own tests.
