+++ camille/__init__.py
@@ -2,10 +2,4 @@
 
 """
 
-from __future__ import absolute_import
-
-from .processors import fft
-from .processors import mooring_fatigue
-from .processors import rolling
-
 __version__ = '0.0.1'
