+++ camille/process/rolling.py
