+++ camille/util/__init__.py
@@ -1,0 +1,1 @@
+from .sncurves import sn_curve
