+++ camille/util/sncurves.py
